+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - skip adding ignored non read-only files to the context
 - improved MCP client interruption handling
 - properly adding user input messages to the input history when using MCP tools
+- wrapping long tool message content
 
 ## [0.3.2]